Use open graph adapter in Link#title
This is the first time the entire test suite is passing, but there is a
still a problem. The test suite now requires a connection to the
Internet to run. If you disconnect, the test suite fails. Additionally
making real web requests in tests drastically slows test the time it
takes to run the suite. We can do better...