Add app configuration for open graph
Now the application itself can be configured with which open graph
adapter to use. This allows us to configure the fake to be used for the
test suite.

The benefit of this over stubbing the OpenGraphAdapter constant is no code
reads as is it's going to behave in one way and later acts in a
different way. We avoid action at a distance.