We no longer need to use the injection interface for model test
Since the test suite is configured to use the fake, we don't have to
explicitly inject it in the model test anymore. This feels OK, because
all code is self-documenting. However, leaving the injectable interface
is convenient as it keeps the implementation of Link decoupled from a
particular open graph implementation.