Implement a thin wrapper around OpenGraph
This establishes an interface that we own, and thus we can mock in the
form of a test fake to decouple our test suite from the Internet.

Note: This test only validates our assumptions about how the adapter
will interacts with the underlying library. It does not verify that
these assumptions are correctly. In fact, you will see that our test
suite is entirely decoupled from the actual opengraph_parser library
which may indicate a lack of test coverage.

See
https://github.com/testdouble/contributing-tests/wiki/Don't-mock-what-you-don't-own