Illustrate one more problem with a sneaky typo
When using mocks and stubs, you always run the risk of accidentally
specifying everything exactly wrong. The tests will pass (these do), but
the actual app breaks. This is because the actual object used when
running the app is not involved in the tests at all. We just assume
(incorrectly) that we got the stubs setup correctly. Whoops.