Make tests fail with RSpec's verifying doubles
RSpec provides us with a nice tool to solve this problem. The so-called
"verifying doubles" verify that messages passed to your double are valid
messages for the class given. Do note, however, that this is a mere
benefit of RSpec. You could accomplish the same thing with any testing
library by asserting each message sent against an instance of the real
object. It might take a little more rigor, but it works.